хуйни почистил r3 нахуй идёт куникор отдельно от псевдовуникса перехуярил ассемблер как любят говорить это фаундейшн ту билд апон ин тхе футуре
писал в час ночи так что без претензий
