+++ file.asm
@@ -1,0 +1,4 @@
+hello:
+  add r4 1
+  hlt
+
