+++ kernel/main.c
@@ -1,11 +1,12 @@
 #include <stdio.h>
 #include <unistd.h>
 #include <stdlib.h>
+#include <string.h>
 #include "init_kernel.h"
 #include "shell.h"
 
 int main(int argc, char** argv) {
-	if(argc < 1) return -1;
+	if (argc < 1) return -1;
 	init_kernel(argc, argv);
 	shell();
 	puts("System will power off right now");
