+++ mall.c
@@ -7,9 +7,10 @@ void compile(char* command, char* name) {
 }
 
 int main() {
-  compile("gcc -pedantic -Wextra -Wall -o mall mall.c", "mall");
-  compile("gcc -pedantic -Wextra -Wall -o vmvuniz-vunix kernel/main.c kernel/init_kernel.c kernel/panic.c kernel/shell.c kernel/vfs.c kernel/rescue.c kernel/exec.c kernel/v.out.c", "kernel");
-  compile("gcc -pedantic -Wextra -Wall -o create_disk create_disk.c", "create_disk");
-  compile("gcc -o writevfs writevfs.c", "writevfs");
+  compile("gcc -Wextra -Wall -o mall mall.c", "mall");
+  compile("gcc -Wextra -Wall -o vmvuniz-vunix kernel/main.c kernel/init_kernel.c kernel/panic.c kernel/shell.c kernel/vfs.c kernel/rescue.c kernel/v.out.c", "kernel");
+  compile("gcc -Wextra -Wall -o create_disk create_disk.c", "create_disk");
+  compile("gcc writevfs.c -o writevfs", "writevfs");
+  compile("gcc exec.c -o vunicore", "vunicore");
   return 0;
 }
