+++ t.txt
@@ -1,0 +1,6 @@
+add r1, 4
+((0,"add"),(1,1),(2,4))
+#add-rri
+((0,"add"),(1,1),(1,0),(2,4))
+0x01100004
+
